[topo] inline id module into root
